Stop four @bb/app suites leaking state into whatever runs next
Vitest orders test files and the tests inside them deterministically, so
these passed every CI run while being order-dependent the whole time. A
rename or a new test file was enough to surface them in an unrelated pull
request. Running the unmodified suite with `--sequence.shuffle` reproduces
all four on main.

Four files, four unrelated causes:

- QueuedMessagesList: `drags the zero-height group handle to a measured row
  stroke` drives a complete pointer gesture through @dnd-kit, which leaves
  state behind that outlives the component and that dnd-kit exposes no way
  to reset. Skipping just that test made the other 36 pass. Moved it to its
  own file, since vitest already isolates per file. This one contains the
  problem rather than fixing it, and says so.

- PluginPendingInteractionComposer: a crashed slot instance is recorded in
  a module-level Set in PluginSlotMount, so a renderer that throws in one
  test disabled that plugin/slot pair for every later test. Wired up
  `resetAllCrashedPluginSlotsForTest`, which already existed and was never
  called. Also `vi.clearAllMocks` -> `vi.restoreAllMocks`: the former only
  drops recorded calls, leaving the file's `console` spies installed.

- SecondaryPanelLayout: `installAnimationFrameQueue` overwrites
  `requestAnimationFrame` with `Object.defineProperty` and then spies on
  it, so `vi.restoreAllMocks` restores the spy to that overwrite instead of
  the real function. Every test after any test using that helper ran with a
  `requestAnimationFrame` that silently never fired its callback — far
  wider than the two tests that happened to fail. Captures the pristine
  functions at module load and restores them in `afterEach`.

- PluginPanelRightPanelHost: `atomWithStorage(..., { getOnInit: true })`
  inside an `atomFamily` reads storage once at atom creation, and the
  family caches that atom for the lifetime of the module. A test that seeds
  storage bakes its value into the atom's initial state, so neither the
  existing `localStorage.clear()` nor the per-test jotai store undoes it.
  Adds `resetFixedPanelTabsStateForTest` to evict the family.

No test was deleted and no assertion relaxed; each change removes leaked
state rather than softening what is checked.

Verified: the full @bb/app suite passes under `--sequence.shuffle` on
seeds 5, 7, 11, 22, 33 and 99 (361 files, 2864 tests), each of which
previously failed or is adjacent to one that did; every touched file also
passes standalone across 6-10 seeds. typecheck, lint and prettier clean.
